little improvements

+++ FileDifferenceForensicEvidence/ge.py
@@ -26,8 +26,8 @@ def execute(configuration: Configuration):
         else:
             return action[1] + "." + str(iteration_number)
                
-    def create_trace_image(action, iteration_number:int):
-        shared_utilities.start_program(configuration,configuration.vboxmanage_executable,"clonemedium disk " + shared_utilities.get_hdd_uuid(configuration,configuration.name_of_vm_to_analyse) + " --format RAW " + configuration.shared_folder_on_host_for_sharing_files_with_vm_which_has_idifference + to_action_name_string(action,iteration_number) + ".raw",0,"Clone medium (Create raw-file with traces for " + action[1] + ")")
+    def create_trace_image(action, iteration_number:int, name:str):
+        shared_utilities.start_program(configuration,configuration.vboxmanage_executable,"clonemedium disk " + shared_utilities.get_hdd_uuid(configuration,configuration.name_of_vm_to_analyse) + " --format RAW " + name,0,"Clone medium (Create raw-file with traces for " + action[1] + ")")
 
     def restore_original_image():
         shared_utilities.start_program(configuration,configuration.vboxmanage_executable, "snapshot " + configuration.name_of_vm_to_analyse + " restore " + configuration.snapshot_name_for_initial_state_of_vm_to_analyse,5,"Restore original state of vm")
@@ -54,36 +54,37 @@ def execute(configuration: Configuration):
         shared_utilities.save_state_of_vm(configuration.name_of_vm_to_analyse,configuration)
 
     def generate_evidence(action,iteration_number:int):
-        configuration.log.info("Start evidence generation for action " + action[1] + " in iteration " + str(iteration_number))
-        try:
-            prepare_generate_evidence()
-            shared_utilities.continue_vm(configuration)
-            if (action[0].lower().startswith("Special:".lower())):
-                if(action[0].lower().startswith("Special:WaitUntilUserContinues:".lower())):
-                    input("Wait for execution of manual action " + action[1] + " ('" + action[0].split(":")[2] + "') in the vm. Please press enter if this action is finished to continue generating evidences.")
-                elif action[0].lower().startswith("Special:Noise:".lower()):
-                    time.sleep(configuration.noise_recording_time_in_seconds)
+        result_file_name=configuration.shared_folder_on_host_for_sharing_files_with_vm_which_has_idifference + to_action_name_string(action,iteration_number) + ".raw"
+        if(configuration.overwrite_existing_files_and_snapshots or not os.path.exists(result_file_name)):
+            configuration.log.info("Start evidence generation for action " + action[1] + " in iteration " + str(iteration_number))
+            try:
+                prepare_generate_evidence()
+                shared_utilities.continue_vm(configuration)
+                if (action[0].lower().startswith("Special:".lower())):
+                    if(action[0].lower().startswith("Special:WaitUntilUserContinues:".lower())):
+                        input("Wait for execution of manual action " + action[1] + " ('" + action[0].split(":")[2] + "') in the vm. Please press enter if this action is finished to continue generating evidences.")
+                    elif action[0].lower().startswith("Special:Noise:".lower()):
+                        time.sleep(configuration.noise_recording_time_in_seconds)
+                    else:
+                        raise Exception("Unknown action") 
                 else:
-                    raise Exception("Unknown action") 
-            else:
-                shared_utilities.execute_action_in_vm(action, configuration)
-
-            shared_utilities.save_state_of_vm(configuration.name_of_vm_to_analyse, configuration)
-            if(configuration.create_snapshots_after_action_execution):
-                snapshot_name = configuration.name_of_snapshots_of_actions + "_" + action[1] + "_" + str(iteration_number)
-                if (configuration.overwrite_existing_files_and_snapshots):
-                    shared_utilities.ensure_snapshot_does_not_exist(configuration,configuration.name_of_vm_to_analyse, snapshot_name)
-                shared_utilities.create_snapshot(configuration,configuration.name_of_vm_to_analyse, snapshot_name)
-            create_trace_image(action,iteration_number)
-            restore_original_image()
-            execute_idifference_for_action(action,iteration_number)
-            delete_trace_image_if_desired(action,iteration_number)
-        except Exception as exception_object:
-            configuration.log.error("Exception occurred while generating evidence  for action " + action[1] + " in iteration " + str(iteration_number) + ":")
-            configuration.log.error(exception_object, exc_info=True)
-        finally:
-            finalize_generate_evidence()
-        configuration.log.info("Evidence generation for action " + action[1] + " in iteration " + str(iteration_number) + " finished")
+                    shared_utilities.execute_action_in_vm(action, configuration)
+                shared_utilities.save_state_of_vm(configuration.name_of_vm_to_analyse, configuration)
+                if(configuration.create_snapshots_after_action_execution):
+                    snapshot_name = configuration.name_of_snapshots_of_actions + "_" + action[1] + "_" + str(iteration_number)
+                    if (configuration.overwrite_existing_files_and_snapshots):
+                        shared_utilities.ensure_snapshot_does_not_exist(configuration,configuration.name_of_vm_to_analyse, snapshot_name)
+                    shared_utilities.create_snapshot(configuration,configuration.name_of_vm_to_analyse, snapshot_name)
+                create_trace_image(action,iteration_number,result_file_name)
+                restore_original_image()
+                execute_idifference_for_action(action,iteration_number)
+                delete_trace_image_if_desired(action,iteration_number)
+            except Exception as exception_object:
+                configuration.log.error("Exception occurred while generating evidence  for action " + action[1] + " in iteration " + str(iteration_number) + ":")
+                configuration.log.error(exception_object, exc_info=True)
+            finally:
+                finalize_generate_evidence()
+            configuration.log.info("Evidence generation for action " + action[1] + " in iteration " + str(iteration_number) + " finished")
 
     def generate_evidence_full():
         for action in configuration.actions:
@@ -93,7 +94,7 @@ def execute(configuration: Configuration):
     def generate_new_init_raw_file():
         restore_original_image()
         shared_utilities.start_program(configuration,configuration.vboxmanage_executable,"clonemedium disk " + shared_utilities.get_hdd_uuid(configuration, configuration.name_of_vm_to_analyse) + " --format RAW " + init_raw_file, "Clone medium (Create raw-file of initial state)")
-        if os.path.isfile(init_raw_file_on_host_for_sharing_files_with_vm_which_has_idifference) and configuration.overwrite_existing_files_and_snapshots:
+        if os.path.exists(init_raw_file_on_host_for_sharing_files_with_vm_which_has_idifference) and configuration.overwrite_existing_files_and_snapshots:
             os.remove(init_raw_file_on_host_for_sharing_files_with_vm_which_has_idifference)
         shutil.copy(init_raw_file, init_raw_file_on_host_for_sharing_files_with_vm_which_has_idifference)
